Introduce git-subrepo script
Including source code managed in other git repositories is a pain. There
are two approaches: submodules and subtrees. Submodules are cumbersome
to manage because a lot of management commands are necessary. They also
do not actually embed the source code of the submodule into the
repository where the module is to be used but merely store a reference
to the hash (with all sorts of disadvantages associated with it).
Subtrees came to the rescue by actually embedding the source code into
the repository. Yet, they have other, more subtle, drawbacks. Once a
subtree is imported (happening via a merge) in a true subdirectory
("prefix") one cannot rebase over this commit without merge conflicts.
The reason is that git forgets about the "prefix" and subsequently tries
to merge the commit in the root. One can try using rebase
--preserve-merges but that results in even more trouble: commits
silently disappear (!). See the BUGS section in the git-rebase (1)
manual.
This change introduces a new mechanism for including other git
repositories: 'subrepos'. Just like subtrees, subrepos embed the other
repository's source code into the current git repository. However,
inclusion does not happen in the form of a merge but as a simple commit
adding the remote repository's files. Hence, rebasing over this commit
is not special in any way and will not cause any conflicts. When
updating the subrepo, the most intuitive thing happens: the difference
between the addition/last update of the subrepo to the commit to import
is taken and applied in the form of a single commit. Again, no merge
commit is created.