Correctly resolve commit names
When importing a subrepo and specifying a SHA1 hash of a commit instead
of a symbolic name (of a branch or tag, for example), git subrepo will
try to resolve this commit by invoking git's rev-parse command. It does
so by supplying the remote repository followed by the commit of interest
(separated by a slash) to the command. This syntax, however, is only
valid for symbolic names and so the invocation will fail when a SHA1
hash is supplied directly.
This change detects such failures and causes git subrepo to check if
indeed a SHA1 sum was supplied. If that is the case, this hash will be
used directly. Otherwise, an error will be reported to the user because
the commit could not be interpreted properly.