Interpret subrepo prefixes relative to the current working directory
Until now a prefix used when adding or updating a subrepo was always
treated relative to the repository's root directory. This is counter-
intuitive and clashes with the behavior of several git commands that
work in a relative manner.
This change causes prefixes to be treated relative to the current
directory. Because internally git works relative to the repository's
root directory, some path magic is used to calculate paths relative to
the root.