Add proper documentation in the form of man, xml, and html files
When asking for the help for a particular git command, usually the man
page of this particular command is looked up and opened. Since for
git-subrepo no such manual is available, behavior diverges from that of
other git commands, resulting in the entire git-subrepo command to be
recognized as somewhat alien.
This change adds a manual page for git-subtree. The man page is
generated from an ordinary text file in the same manner git's own manual
pages are created. Furthermore, since git also delivers xml and html
files generated from the text file, such files are created for
git-subtree as well.