Use execute module for process execution
Until now we used the subprocess module for starting new processes (in
order to run commands). With an upcoming change we will require
pipelining functionality (meaning the ability to directly supply one
process' output as another one's input) which is not provided by this
module.
To that end, this change converts the program (and the tests) to use the
custom execute Python module. This module comes with support for
pipelining and the conversion is straight forward.