Add support for importing into the repository's root directory
One advantage of subrepos over git's native submodules is the fact that
the source code of the destination is embedded in the repository.
However, that such embedding not only comes with benefits but also with
disadvantages. One of those drawbacks is that the level of nesting of
subrepos can increase extremely when each repository pulls in its
dependencies. Consider the following example comprising two
repositories:

  lib1
  └── src
      └── lib1.py

  lib2
  └── src
      ├── lib1
      │   └── src
      │       └── lib1.py
      └── lib2.py

In this example, 'lib2' depends on 'lib1' and, hence, embedds it in the
form of a subrepo. It does so in its 'src' repository (but that is not
relevant for the illustration of the problems). Next imagine an
application that depends on 'lib2' but also on 'lib1'. When employing a
similar directory structure, the repository will look like the
following:

  app
  └── src
      ├── app.py
      └── lib2
          └── src
              ├── lib1
              │   └── src
              │       └── lib1.py
              └── lib2.py

There are two problems with this scheme: First, as mentioned before, the
level of nesting increases with each dependency. If 'app' was imported
in another repository the level would increase again. Second, and
slightly related is a non-uniformity in terms of the directory layout
(source code of subrepos cannot be located at the same nesting depth as
that of the parent repository) that is unavoidable.

This change addresses this problem by introducing the necessary means to
import a subrepo within the root directory of the current repository.
Using this feature, there is no strict requirement for an ever
increasing nesting depth if repositories are organized in a way that all
subrepos are embedded in the respective root directory. In the example
given earlier 'lib2' could look like (after adjusting 'lib1' and 'lib2'
by inserting an additional directory):

  lib2
  ├── lib1
  │   └── src
  │       └── lib1.py
  └── lib2
      └── src
          └── lib2.py

Such a structure was impossible previously. The reason for this
restriction was that the command used for embedding the subrepo,
git-read-tree, treated a prefix being the repository's root directory in
a special way in that it attempted a merge that could easily conflict
with other files.
Because of that peculiarity, the implementation was adjusted to rather
use a combination of git-diff-tree and git-apply which does not impose
such a constraint (although, arguably, at the expense of being less
elegant from an implementation stand point).
Note that the root repository's root directory acts as a shared
namespace. That is, multiple subrepos that have equally named
directories will cause a clash. Subsequent changes will alleviate this
problem.