Allow "merging" of similar subrepos imported into a repository's root
A previous change motivated why it can make sense to import subrepos
into a repository's root instead of into a dedicated prefix directory
and enabled such usage. One problem that remained, however, was the
potential for conclicts when multiple subrepos have in turn dependencies
to the same remote repository.
For instance (expanding on the previous example), consider that we have
three libraries: 'lib1', 'lib2', and 'lib3'. 'lib2' and 'lib3' each
depend on 'lib1' (and embed it in their root as a subrepo). Graphically:

  lib1
  └── lib1
      └── src
          └── lib1.py

  lib2
  ├── lib1
  │   └── src
  │       └── lib1.py
  └── lib2
      └── src
          └── lib2.py

  lib3
  ├── lib1
  │   └── src
  │       └── lib1.py
  └── lib3
      └── src
          └── lib3.py

Now imagine an application, 'app', requires both 'lib2' and 'lib3' and
wants to import them as subrepos into the project's root. Since each
library depends on 'lib1' (which will be pulled in accordingly), a
conflict will occur because 'lib1' is to be imported twice with the same
name.
This change addresses this issue by respecting files and, more
importantly, directories that already exist in a repository when
importing a subrepo. The idea is to apply the changes that are required
to get a subrepo to the desired state *on-top* of any such existing
files. With this approach, we can get to the following state:

  app
  ├── app
  │   └── src
  │       └── app.py
  ├── lib1
  │   └── src
  │       └── lib1.py
  ├── lib2
  │   └── src
  │       └── lib2.py
  └── lib3
      └── src
          └── lib3.py

The rationale behind merging the two potentially different states of
'lib1' is the following: if we imported 'lib1' in two states (one as it
was imported by 'lib2' and the other as it was imported by 'lib3')
depending on the language/toolchain/interpreter we might not be able to
actually make use of both of them anyway. Consider Python: if we ran
'app' we would have to embed both 'lib1' versions in the PYTHONPATH.
Yet, only the first one that is found will be used throughout the run
time (resulting in the other one effectively being dead code). Compiled
languages behave similarly (though not necessarily equally). In C/C++
one would have to set the include path and the library path (if needed).
In both cases the first header or library matching the name searched for
will be used.
Hence, both versions of the subrepo need to be "compatible" if they are
to be used in a common application and we must be able to agree on using
a single one.