Remove distinction between 'add' and 'update'
With recent changes the need to distinguish between the 'add' and
'update' commands vanished. We always update a subrepo to the desired
state and internally there is no need to differentiate between an
initial import and an update. More so, once a common subrepo (call it
"foobar") is pulled in from multiple other subrepos (because they depend
on it; call them "foo" and "bar", respectively), the distinction is void
anyway: although there is an initial 'add' happening when the first
subrepo (say, "foo") pulls in "foobar" as well, the import of "bar" is
an addition from the point of view of "bar" but not of "foobar" because
the latter now already exists.
To that end, this change only exposes a single command, 'import', that
pulls in a remote repository at a specific state in the form of a
subrepo.
As an aside, the concept of an 'update' never really was that because
strictly speaking one could downgrade as well by specifying a state
created earlier than the currently imported one. This behavior is
preserved.