Report that no changes were made when importing at already present state
Nothing prohibits users from importing a subrepo that is already
imported. In fact, that is the way to update or downgrade it. However,
it is also possible to import it at the state that is already present,
in which case effectively no changes will be made.
Until now, git-subrepo supported this case (in the sense that no
unexpected side-effects would happen) but reported a user unfriendly
message back to the invoker because the git-commit operation would fail.
This change adds a check to detect if no changes were made and reports
an error in such a case.