Report that no changes were made on importing an empty repository
It is possible to import arbitrary remote repositories as subrepos,
including those that effectively are empty (because they only contain an
empty commit). In such a case we want to correctly report that no
changes were made back to the user as we do for the case of importing a
subrepo at the already present state.
This patch adds support those scenarios.