Do not print backtrace in case of a ProcessError
In case one of the git commands fails we retrieve a pretty verbose
backtrace. Although the information conveyed in it potentially helps in
debugging issues the error is very likely due to erronous input by the
user. In such a case having a high level of verbosity just makes it
harder to understand the actual problem quickly.
This change addresses the problem by only printing the output as printed
by git.