Introduce -d/--debug option
With the previous change to not print a fully-fledged backtrace but just
the condensed error by default we might be restricted access to
important information for investigating a problem.
This change introduces a new option, -d/--debug, to get back the
previous behavior of printing a backtrace for any error.