Error out on chached changes or conflicting unstaged files
When importing a remote repository using git-subrepo unstaged and staged
files alike would simply be discarded, potentially throwing away
precious work.
We want to play safe and not take chances throwing away anything useful
and so this change adds the necessary checks to error out in case of
staged changes or unstaged and conflicting changes (i.e., files
contained in the working tree but not the index that would be
overwritten by an import).