Properly export functions relevant for importing a subrepo
Other Python packages might want to make use of git-subrepo without
having to start a new interpreter instance for doing so. To that end,
this change exports the most relevant functions for an import into the
git.subrepo package. A slight amount of refactoring was required to
expose the same parameters as required in the command line interface:
the remote repository name, the prefix to import at, and the commit
representing the state at which to import.