Fix failure reimporting from a directory other than the root
When importing a subrepo that is already contained in some different
state from a directory other than the repository's root, the import will
fail. The problem is that the git-apply operation works relative to the
current working directory and so applying the patch fails because the
existing contents do not match. One possible way around this problem
would be to use the correct strip value, but there is be more to that
and complexity is increased unnecessarily; we have three variables: the
prefix [relative to *some* directory], the current working directory,
and the repository root.
This change takes a different approach: (almost) all git invocations are
invoked with the -C option pointing to the repository's root. This way,
git will implicitly change into the correct directory before performing
any operation, eliminating the current working directory from the
equation completely.