Check that a commit belongs to a given remote repository
We currently unconditionally request the user to specify the remote
repository which to import in the form of a subrepo as well as the
commit in it representing its state. This approach seems fine and it is
unclear how one would go about deducing affiliation of a commit to a
remote repository in case it is not a symbolic name but a fully
qualifying SHA1 hash. One problem, however, can be that for various
reasons there is a mismatch between the commit and the remote repository
that it supposedly belongs to.
This change adds a check to the import script that verifies that the
commit does indeed belong to the given remote repository. If that is not
the case, the import will be aborted. The check only allows only
commits reachable to be specified (and no loose ones). There could
potentially be cases where the user wants to circumvent this check and
so a new option, -f/--force, is added that does precisely that.