Rework subrepo import
Our approach of importing by means of simple patch arithmetics works for
most imports but there are cases where we leave stale files when
performing a root-level import. Consider the case where files in the
remote repository were renamed between the last import and the desired
new state. Upon importing the new state, the resulting diff would only
contain the additions over the last import, not the removals. As it
turns out this problem is inherent in the approach taken (and only
affects root imports).
To fix this issue two potential solutions were investigated:
1) We need to find the commit of the previous import of the given remote
   repository at the given prefix. We can then create an incremental
   patch that only contains the changes since the last import without
   performing any patch "calculations".
2) We revert the contents of all files and directories imported with the
   current commit that potentially were also imported earlier. This way,
   we import the "new" state on top of a pristine environment (with
   respect to any previous imports), thus not leaving any stale files
   behind.

At first glance, the first approach seems more reasonable than the
second. However, there are subtle issues to it that render it
unfeasible.
The problems start with finding the previous import commit. We do not
want to keep track of imports in a file (say, .gitsubrepos) so the only
viable option is to grep the history for imports that match a certain
pattern (the possibility of using notes to contain import meta data was
discarded as it turned out that notes are not retrieved by default when
fetching contents of a remote repository). When searching the history
for a certain pattern, not only are we prone to false-negatives if the
user tinkered with the commit message, we potentially incur a high
runtime cost because we have to trawl a potentially long history for a
matching pattern. On top of that, sifting through the importing
repository's history is not enough. Two different remote repositories
can import the same remote repository by themselves. For example,
consider the following repository hierarchy:

      app
      / \
     /   \
  lib2   lib3
     \   /
      \ /
      lib1

If we import 'lib2' (with an up-to-date version of 'lib1') into 'app',
rename a file in 'lib1', and then update 'lib3' (again, including the
most recent version of 'lib1'), we can still miss deletion of the source
file of the rename operation in 'lib1'.
Hence, we need to search the imported repository for *all* subrepo
imports (which also means we need to embed each of them in the
respective commit message as well as introduce a "delete" command,
otherwise a subrepo will be recognized as such until the end of time,
even if the source code effectively got removed).
With this suggested architecture we are able to deduce the imported
commit of a subrepo import. However, there is no guarantee we can
actually work with it. If, as in the above example, 'lib1' is only known
indirectly to 'app' (because it is a dependency of 'lib2' and 'lib3',
'app' itself has no knowledge of any commits in 'lib1'. So creating a
diff against a previous state of it is impossible.

Long story short, this patch uses the second technique to solve the
aforementioned issue. We look at the files and directories in the commit
and remove all of them before importing the desired state. This approach
still leaves open the possibility of having renames of the (top-level)
directories. Such renamings are considered very unlikely in the case of
root-level imports (they are somewhere on the level of project renames
because the entire source code should reside below such a directory) but
nevertheless will be subject to subsequent changes. Using the second
approach also has the advantage of reducing code complexity since we no
longer require two largely separate paths for root-level and prefixed
imports.