Correctly handle renames in the root directory in the general case
With an earlier change we pushed the renaming problem (that is, the
problem that -- if administering a root-level import -- file renames in
a remote repository when performed between two subrepo imports of the
same can leave stale files in the importing repository) to affect only
files and directories in the root directory.
This change goes one step further and shrinks the problem space even
more: If the remote repository is imported directly, i.e., not as part
of an import of another remote repository, we are able to mitigate the
problem by removing not only the files and directories as comprised in
the current state to import but also those from the previous import. In
order to retrieve this commit we have to search the commit history for
messages matching the pattern of an import. We worked very hard
previously in order to not have to fall back to such a search and
pattern match but it seems to be the only viable approach.