Correctly handle renames in all cases
A previous change used the history to infer the most recent (if any)
import commit of the same repository being imported. Knowledge of this
commit helps to properly cleanup potentially stale files in case of
rename operations in the imported repository. The approach works for
most imports but it fails if a repository is pulled in through a remote
repository but is also embedded directly into the importing repository.
For instance:

  app
   |\
   | \
   | lib2
   | /
   |/
  lib1

In this case, if 'lib1' is updated independently of 'lib2' with a file
rename having taken place between the last import and the new one, a
stale file will be the result.
This change addresses the problem in a comprehensive manner. Whenever we
perform an import we not only remember the state of the imported
repository we are importing, but also the state of all other subrepos
that were imported in this (remote) repository in the past. As before,
we do so by means of the history, i.e., by embedding the SHA1 sums of
the commits of interest into the respective import commit. Now, because
of this mechanism, we can detect implicit updates of subrepos (that is,
imports of subrepos which in turn contain other subrepos) by scanning
the log.
In the example involving 'app', 'lib2', and 'lib1' above, an update of
'lib2' in 'app' would implicitly update 'lib1' and would remember it as
such. Hence, a subsequent explicit update of 'lib1' could reference the
state of the other, implicit update that just happened before.
From a user's point of view, this change is beneficial as well because
the log now states a subrepo's dependencies much more clearly (by having
them all listed in the commit message).
On the downside, we increase the complexity (in terms of runtime cost)
for an import. Whenever an import is performed, we need to scan the
history of the imported as well as the importing repository -- a
potentially costly operation. The scan happens in the form of a "grep",
however, and experiments have shown that even on a repository with a
commit count north of 100k such a pattern search is still performed in
under six seconds. Hence, the duration for an import is dwarfed by the
time it takes to even add the remote repository and fetch its contents
in the first place. Plus, we do not expect imports of repositories of
that size.