Introduce GitImporter class
Over the course of development, two rather unpleasant constructs were
introduced into the architecture: First, in order to increase verbosity
of executed commands we used a global variable. Such global state would
not be that bad if we only supported the invocation as a script (that
is, the direct execution in a Python interpreter), but we also allow
using the import_ function (and others) from another script that just
imported this functionality. Second, we pass the repository's root
directory to every function that directly or indirectly executes a git
command.
Both issues exist because we have some state which is shared in between
function invocations -- a scenario well suited for a more object
oriented solution. To this end, this change refactors the code by
introducing two classes. The GitExecutor class is used for encapsulating
this state required in essentially every git process invocation. Based
on an object of this class, GitImporter provides the remaining
functionality for performing imports.