Do not pass around remote repository imports
Returning a dictionary of imports of the imported repository from the
import_ method is a lousy way of avoiding a costly search through the
imported repository's history from two places.
This change uses a cache in order to remember the return value of
_searchImportedSubrepos which aids in not paying the cost for a repeated
invocation while not pushing the burden of temporarily storing the
return value to the user. By doing so we also fix a bug in the imports_
function where its return value was not the correct set of imports in
case the importing repository did not have any commits yet.