Use the correct prefix for commit messages in nested imports
When importing nested a repository, i.e., a repository that contains
another repository in the form of a subrepo, the resulting commit
message's prefix does not necessarily reflect the actual import prefix
of the repository.
This patch corrects the prefix appearing in the commit by properly
joining the prefix of the nested subrepo with prefix of the current
import.