Introduce Subrepo class
A subrepo is always represented in the same way: as a (repo, prefix)
tuple.
This change introduces a new class, Subrepo, that represents such a
subrepo in a more expressive and consistent way. Objects of this class
can automatically be converted into a string of the known form
"prefix:repo".