Introduce subrepo deletion functionality
Recent changes causing the subrepo import to look up all previous
imports make it a necessity to be able to remove a previously imported
subrepo. This requirement stems from the fact that when creating a new
import commit for a certain remote repository we embed the latest state
of all this repository's imports into the commit message (which gives a
nice idea of the dependencies pulled in by this import). Naturally, we
might want to remove a subrepo at some point but even if the code got
removed manually, the history would still show an import and, hence,
lead to a wrong commit message.
This change introduces a new subrepo command, 'delete', in addition to
the already existing 'import'. A subrepo deletion is marked in the
history as such and, hence, the import code can find (and ignore)
already deleted subrepos.