Add 'tree' command
The existing git-subrepo commands build a dependency tree for all the
imported subrepos before performing any action. On occasion, it might
come in handy to visualize this tree in order to get a better impression
of the sub-components of a repository.
This patch introduces a new command, 'tree', that can be used to
visualize the dependency tree by printing it to the standard out stream.