Add 'reimport' command
When working with subrepos it can happen that a subrepo's history needs
to be amended, be it to include a small fix that slipped during
development or the correction of a typo. In such a case the subrepo
should be updated at some point. This procedure is a tedious task
because commits have to be looked up and matched manually and a reimport
has to be performed that also includes adjusting the original commit
message.
This task can be automated and this is what the newly introduced
reimport command is made for. The command is to be used in rebase
operations. It inspects the commit message of the current HEAD to see if
it is a subrepo import. If that is the case, it extracts the imported
commit hash. Based on this hash it extracts the subject line and matches
this subject against "newer" commits. If it finds one, it will amend the
commit to import the most recent state, while adjusting the commit
message accordingly.