Add support for imports including additional changes
Sometimes changes to a repository are not compatibility preserving,
meaning that clients have to be adjusted to conform with them. If a
repository containing such changes is imported as a subrepo, it is
necessary to not only perform the import to get the new state but to
also perform some changes to local client code in order to make it work
with the newly imported state. For the sake of atomicity, these changes
should be included in the import commit.
This change adjusts the git-reimport script to handle such commits along
with those "just" importing a subrepo. Also, in such a scenario the
commit message would not be solely the import statement but rather
contain some additional text.