Introduce -b/--branch parameter
git-reimport searches commits in a remote repository based on the
subject line. If there are multiple branches with one having been
rebased recently and, thus, having different commits, the script will
detect that and not know based on which of those commits to reimport.
The user, however, should know from which branch contains the most
up-to-date commit.
This change introduces the -b/--branch parameter that can be used to
restrict search for a commit to a certain branch in a remote repository.
This way, such ambiguities can be avoided or resolved with basic
knowledge.