Rename --debug to --debug-exceptions
The -v/--verbose option, although increasing verbosity, is really a
means for better debugging: There is not much of a point to use it under
normal circumstances.
To that end, an upcoming change will rename it to convey this fact more
clearly and to not reserve the --verbose option which really lends
itself more towards verbosity increases that are less debug related. In
preparation for that change this one renames the -d/--debug option to
--debug-exceptions, in order to allow for multiple debug related options
to coexist.