Introduce -v/--verbose option for reimport
It can be of interest to see what a reimport actually does for a
particular commit when performing a rebase over a reasonably large
commit history.
This change introduces a new option, -v/--verbose, that, when used,
causes the command to print some more status messages about whether or
not a reimport is happening, and which commit from the remote repository
is used.