Adjust reimport command to handle multiple imports properly
When we perform a reimport we need to adjust the commit message to
reflect the new state being imported. With this patch we can handle
multiple imports (i.e., a direct one and multiple indirect ones as part
of it) correctly. Before this change we assumed only a single import
and, hence, simply performed a replace of the old import SHA1 hash with
the new one.