Enhance reimport command with deletion support
As of now the reimport command did exactly what the name suggests:
reimport an import commit. However, with the introduction of deletions
it is important to reimport subrepo deletions as well. Although
generally there will be less of a need to adjust the message (only if
additional dependencies got deleted as part of the deletion), it can
happen that due to a reimport new files got pulled in earlier which now
will also have to be deleted.
This change adjusts the reimport code to add support for detection and
reexecuting subrepo deletions.