Rework sub-command dispatching
The dispatching of sub-commands has always been a bit messy, with string
comparisons that are prone to become outdated once changes are made but
also the existence of some attributes only conditionally for some
sub-commands.
This change reworks this dispatching and gets rid of these problems. We
use the subparser's set_defaults method to unconditionally set the
'perform_command' attribute which can be invoked uniformly by callers
(because the signature is the same among the different commands).