Correctly import files residing in a path that is a prefix of another
Our logic for removing files from a list of files and directories that
were contained in those directories is incorrect. Because the
os.path.commonprefix function which we use for deciding whether a file
resides in a given directory works on the level of strings and not (as
the package affiliation suggests) on individual path components as a
whole, we might remove files from the list of files that were actually
not subsumed by a directory.
This change fixes the problem by reworking the removal algorithm to
detect such cases. Note that the Python community has realized this flaw
of os.path.commonprefix and created a new function, os.path.commonpath
(available in Python 3.5 or higher), that works on the level of path
components. Unfortunately, we want to stay backwards compatible with
older Python versions and so we cannot yet rely on this function.