Fix possibility of non-unique commit detection om reimport
When reimporting a subrepo finding the new commit to import of the
remote repository is determined using pattern matching on the subject
line. Because of wrong parameters being passed to the git command used
for finding this commit, a commit with this subject line contained in
the importing repository (not the imported one) would also be found,
potentially resulting in multiple possible commits to choose from in
which case the reimport is aborted.
This patch fixes the issue by not searching commits reachable from HEAD
but only those in the remote repository of interest.