Fix error when a branch exists that has the name of the import prefix
When a repository contains a branch that has the same name as the
directory we import into, git reports an error because it cannot
unambiguously resolve the second argument to a git-diff-index
invocation.
In order to tell git that we are dealing indeed with a path, prepend two
dashes ("--") before the second parameter to disambiguate it.