Use argcomp package to provide argument completion
As a command line based program, the different parameters that are
understood need to be remembered and typed. Short options alleviate part
of the burden by requiring less typing but that is not sufficient.
Instead, proper tab completion is essential for providing a better user
experience.
To that end, this change makes use of the argcomp package which provides
an argument parser class that can act as a drop-in replacement for the
ArgumentParser class of the argparse module of the standard library. By
doing so we get basic completion of keyword style options.