Add support for context sensitive completions
Out of the box the argcomp package can provide completions for certain
arguments such as keyword arguments or simple choice-based positional
ones. However, some of the arguments of importance in the git-subrepo
program are of a more context sensitive nature and cannot be inferred
without additional input.
This change introduces custom "completer" functions that allow argcomp
to complete bits such as the name of a remote repository or a prefix.