Use context manager for chdir invocation
For our testing we rely on changing the current working directory into a
git repository's sub-directory in some paths. In order to make sure that
the repository can be removed successfully afterwards, we must not stay
in this repository during cleanup and so we revert the working directory
after performing the test. To be exception safe, however, the code is
relatively clumsy.
This change uses a proper context manager to reduce the syntactic sugar
in individual test cases.