Complete prefixes relative to current directory
The prefixes reported by git-subrepo are always relative to the
respective git repository's root directory. However, that is not the
desired behavior we want when completing prefixes for if a user is in a
sub-directory already, the resulting computed prefix would be a wrong
and the import or delete operation would fail.
This change fixes the problem by converting a prefix found to be a
completion candidate to one relative to the current directory.