Hard-code git-subrepo in completion support script
When performing tab completions on git-subrepo in a virtual environment,
the required packages (e.g., deso.argcomp) cannot be located. The reason
for this behavior is that those packages are installed globally and,
hence, are not accessible from inside the virtual environment. This
problem exists with git's completion infrastructure as well in that the
moment we have a commit hook invoking some Python functionality the same
problem arises from within a venv.
This change adjusts the completion script to hard-code 'git-subrepo'
inside the script. The result of this change is that by setting an alias
including the proper PYTHONPATH setting, the 'git-subrepo' string
effectively gets expanded to 'PYTHONPATH="..." git-subrepo'. This is
precisely the behavior we want as now each git-subrepo invocation (and
only it) has access to the system's Python packages.
This path was chosen because it is the way we resolve the same problem
for git. The git completion script hard-codes 'git', resulting in the
same "preprocessing" being available to resolve this issue.