Auto complete existing directories for import command
If an attempt is made to complete the prefix attribute of the 'import'
command, the suggestions include only prefixes where the respective
remote repository has already been imported in. This is fine in
principle but in case the remote repository has not been imported at
all, we would wish for existing directories being suggested in the form
of completions.
This change adds the logic to accomplish this task.