Only perform context sensitive completions inside of a repository
If we attempt to invoke a git-subrepo completion outside of a git
repository we see a long backtrace caused by some invocation of git that
fails to find the repository root. In fact, all context sensitive
completions that rely on some git functionality are not safe to run
outside of an actual repository.
This change introduces a decorator that makes sure we bail out silently
in case such a completion is wanted outside of a git repository.