Check that branch specified via -b/--branch does exist
When performing a reimport and specifying a branch in which to search
for commits, we internally invoke git rev-list with the --remotes option
supplying a pattern containing the branch name. Unfortunately, if that
branch does not exist git rev-list prints out usage information as
opposed to erroring out.
To fix this problem on the side of git-subrepo, this change adds a check
to verify if the repository-branch pair represents a valid commit (and,
thus, verifying the validity of the branch) and raises an error if that
is not the case.