Correctly report invalid commits during import
If upon import the user specifies a symbolic name that can be resolved
locally but not for the remote repository (think, for example, HEAD), we
raise a RuntimeError saying that the commit is not understood. That is
okay in principle but we do not properly catch RuntimeError objects in
the main part of the program and so we print an exception (actually we
print two because we are in an exception handler and the previously
active exception will appear as well). This behavior is undesired but
closer inspection has also shown that the error message is suboptimal in
that it differs between the input of completely unknown commits and
those that have a meaning locally.
To that end, this change reraises the original exception which is
closest to what we expect to be reported.