Use case-insensitive matching of commit messages
The Git community seemingly has agreed on a best-practice style of
commit messages. In particular, the recommendation is to begin commit
messages with an uppercase letter.
git-subrepo currently does not adhere to this style as all produced (and
understood) commit messages start lowercased. To that end, this change
adjusts the program to produce capitalized commit messages and to
understand and recognize commits irrespective of their case convention
(which can help during the transition, but should not hurt even past
that point).