Add support for reimporting based on commit dates
Occasionally commits get reworded slightly, in which case the subject
line may change as well. That can be a problem for the reimport
functionality, because in that case it is no longer able to find a
matching commit to reimport.
To cope with this problem, this change adjust the logic slightly to also
take into consideration commit dates and not only subjects when
searching for the commit to reimport.
As the user may have mocked with commit dates as well it is
theoretically possible that the wrong commit is imported. Because of
that this functionality is not enabled by default but has to be
explicitly asked for by means of the newly introduced --use-date option.